Refactored colvarbias_harmonic to colvarbias_restraint and colvarbias_harmonic and added colvarbias_alb
I've refactored colvarbias_harmonic to be a colvarbias_restraint so
that there may be code re-use between a linear and harmonic
restraint. This also allows me to place the colvarbias_alb, the
adaptive linear bias, into the class hierarchy. This commit also
includes the initial version of the colvarbias_alb implementation.