read and parse index file in binary mode
seek on text files is not reliable on windows due to the
automatic CR-LF conversion. the file stream will not be
repositioned correctly. since the parser itself can handle
files with LF and CR-LF conventions equally well, opening
in binary mode should work, even when reading LF terminated
files on windows and CR-LF files on linux