fix bug due to uninitialized variable in LAMMPS proxy class
This fixes the bug report by Michalis Palaiokostas-Avramidis
from 2014-04-16 on the lammps-users mailing list. When reading
a restart the colvarsRestartFrequency is initialized to the value
of the restart_every member variable. however, this one is only
set, when either a restart command is included in LAMMPS, or
colvarsRestartFrequency is set explicitly.