add support for multi-replica runs in LAMMPS, small refactoring
this change adds fundamental support for using multiple replicas
to the LAMMPS backend. it is only tested with fully independent
calculations. actual multi-walker ABF may need additional programming
to avoid deadlocks as they can happen with plain MPI based
blocking communication, but usually does not apply to charm++.
also, there is a small change to the receiving communication method,
to compare the storage buffer size to the received message size.