Allow setting a feature state while parsing its kewyord
An example is implemented for debugGradients (TODO: get rid of more bool temporaries).

Note 1: the new overloaded get_keyval() will call enable() through the calling deps object.  Is there a case where this may not be wanted?

Note 2: the feature_state class now holds a pointer to the deps object: this is to avoid inheriting colvarparse from deps, but that may be standard behavior in the long run?