Allow getting/setting colvardeps features through colvarscript
This allows controlling object flags at runtime (for example the "active"
flag) for optimization purposes; also allows fixing implementation errors
through the interpreter after deployment.

TODO: the disable() method does not allow dependency resolution yet, and the
list of features is not documented for end-users, so the usefulness is mostly
for the second scenario.