Remove a few more static members, simplify initialization of atom_group
In the colvarparse class, references to dummy strings are replaced by
pointers, and the white_space constant is replaced by a C-string.

The commit also simplifies how the configuration of atom groups is read in the
process (all error checking is now done by the parent object, avoiding
micro-management of the save_delimiters flag).