Add proxy reset method
Upon cvm::reset() e.g. triggered by a script, this method can reset
the proxy's internal state as well as any change to the MD back-end.
Implemented in NAMD with this commit to clear any requested atom and group coordinates.
To be implemented in the other proxies. Not feature-critical, but can
improve scaling in dynamic scripts.