Fix outputCenters parsing
In replica exchange umbrella sampling simulations, it is useful to write out the center of the window to the traj file, so that the traj files alone give you an accounting of exchanges between windows, which used to be enabled by turning on "outputCenters". Since this is not a traditional moving window, currently the code skips parsing outputCenters input, since the return statement comes first. Commenting this out works for me, although I do not know what the TODO is referring to on line 287.