Accompany all section titles in the doc with a label
The current hack introduces right before each section title a horizontal line,
which is labeled and immediately linked.  This allows using permanent URLs
that point immediately before the section title.

Unfortunately, the label of the line is still different from the label of the
section, and the table of contents shows URLs with the automatically generated
labels.