Move OpenMP support to base colvarproxy class
Charm++ 6.8.0 now includes support for thread comanagement with OpenMP.
CkLoop remains, but is very likely to be phased out quickly: its adoption
wasn't very wide in the first place.

The OpenMP-implemented proxy functions, formely restricted to LAMMPS, are now
part of the base proxy class.  The CkLoop-specific NAMD proxy functions remain
available for legacy support of CkLoop, and are only compiled in that case.