Convert the input keyword to lowercase in read_state_data_key
colvarbias::read_state_data_key compares the input keyword from the
state file with the hardcoded one. The hardcoded keyword is always
convert to lowercase but the input keyword is not. As a result, the
following code fails:
(In function write_state_data()):
    os << "grid_dV\n";
(In function read_state_data()):
    if (! read_state_data_key(is, "grid_dV"))
This commit fixes the problem.