Update colvartypes.h
When I submitted this pull request earlier tonight, it seems that I was making edits to the LAMMPS version of this file which is older (instead of starting from the Colvars/colvars repository version).  Fortunately, thew two versions of the file are almost the same, and it was easy to make an update.