Fix hyperlinks
The formatting update to the section headers broke some of the internal
references to labels.  I couldn't figure out how to fix the references to the
parent section in \refkey{}, but since that is only needed for the NAMD and
VMD docs, I removed them (the hyperlink to the keyword works).