Replace sections of VMD and NAMD user guides with links to GitHub Pages
Recent formatting improvements aimed at improving readability have made it
more difficult to remain compatible with the LaTeX environment used when
building the VMD and NAMD user guides.

This commit replaces the (now rather lengthy) Colvars sections of those user
guides with simple links to the Colvars web docs.  This is accompanied by
having added a level corresponding to the package's major version to the doc
webpage tree.  For example the following URL is current:
https://colvars.github.io/vmd-1.9.4/colvars-refman-vmd/colvars-refman-vmd.html