Support the reference group in polar phi and polar theta CVs.
The previous implementation of polar phi and polar theta angles assumed
that the vector to derive spherical phi and theta angles always starts
from the origin (0, 0, 0), which is not always the case. This commit add
an option called "ref", allowing to specify a reference group of atoms
as the origin point, so now we can take a vector from "ref" to "atoms"
and compute the phi and theta angles upon it.