Stop tracking NAMD Make.depends
It has been a while now since the Make.depends file for Colvars was split off
from the NAMD one, so any changes done in the Colvars repository would not
affect it.  Additionally, recent changes from the NAMD tree were not
reflected by an update in the Make.depends file.  For these two reasons, it
is not useful to track it anymore.