Fix segfault when computing max force index
The variable atoms_max_applied_force_id_ was used alternately to
represent an index in the colvars atoms arrays and in the topology.
It was not reset from one timestep to the next, so that translation from
one to the other could happen several times in a row, leading to
nonsense values.
Fixed by adding belt and suspenders:
- use a local variable to represent internal index within atoms_ids[]
- reset atoms_max_applied_force_id_ to -1 if no index is returned (eg.
  if zero forces)
- initialize index within compute_norm2_stats