Hash features when they are reported, and report new ones if needed
Uses std::hash<> (yay!) to track which features were last reported (if any)
and if needed print a new report.

So far limited to NAMD, the reporting mechanism in LAMMPS is different, VMD
still needs to be driven by the Dashboard or by a script, and GROMACS does
not support yet #419.