Load recent toolsets if they are available
GCC 4.8 is a good reference, but won't cut it when we need to build GROMACS.
I updated the Singularity container image to include devtoolset-{7..10}, and
will plan on loading them later.

For now, having a more recent Git than the one shipped with CentOS 7 is a big
help (support for worktrees, among other features).