Split the colvar::init_components_type into C++11 and non-C++11 cases
I end up with splitting colvar::init_components_type into two code
paths, one for C++11 and the other for non-C++11. In the C++11 path we
support nesting CVCs, which requires the lambda functions in the
global_cvc_map to work. But before initializing all components, we
should have all available CVCs in global_cvc_map, so in the C++11 path,
the names and constructors of the CVCs are inserted to the map at first,
and then the map is iterated to initialize the actual CVCs with the
configuration string.