Simplify the calculation of derivative of Q wrt S
The overlapping matrix S is always symmetric, and when we compute the
derivative of its eigenvector or eigenvalue with respect to its
elements, we also expect that the small change of S, dS, to be symmetric.
Let q be the leading eigenvector of S, that means we could assume

dq_k/dS_{ij} = dq_k/dS_{ji}

So we could further simplify the calculation.