fix: try to solve #87 for non-scala components
We have already had the algorithm for computing the gradients of the
fitting group from the gradients of the main group. In cases that the
gradients of the main group are not available, we can "intercept" the
forces applied to the main group, and then use the same algorithm to
convert the forces applied to the main group with the rotation to the
forces applied to the fitting group.