refactor: unify the exertion of fitting group forces
This commit introduces the group_force_object that unifies the exertion
of atomic forces in an atom group. If a colvarcomp requires to apply
forces individually on atoms in an atom group, now it can do
```
auto ag_force = atoms->get_group_force_object();
for (size_t ia = 0; ia < atoms->size(); ia++) {
  const f_ia = ...;
  ag_force.set_atom_force(ia, f_ia);
}
```
If the optimal fitting is enabled for the atom group, group_force_object
will "intercept" the forces applied to the main group, and then calculate
the forces required on the fitting group by calling calc_fit_forces
internally, and then apply the forces on destruction.