Prevent usage of distance functions on incompatible types
Types indicating differences or derivatives of unit vectors and unit
quaternions are needed to handle values with the same dimensionality as the
original types, but no constraints.

The two enum also need to be handled explicitly in a switch statement to
compile without warnings.  However, it is better if the corresponding code
raises an error, instead of computing something that has no consistent
definition.