refactor: using custom allocator for CUDA
This commit uses a custom allocator for the containers of positions,
applied forces, total forces, masses and charges. The custom allocator
can ensure that the vectors are allocated on host-pinned memory, so that
the CUDA transpose kernels can directly transpose and copy the data from
GPU to host, which reduces the data moving.