Backport recent fixes to NAMD 3.0 (#837)
* fix: fix the calculation of bias factor with off-grid hills

When the hill to be deposited is off-grid, we cannot obtain the energy
from the grid as that would be an out-of-bound memory access. This
commit sums all off-grid hills to get the bias energy to compute the
bias factor for WT.

* fix: fix the gradients of the eigenvector CVC

The fit gradients are very significant and cannot be ignored!

* fix: try fixing the orientation sign

An attempt to fix #827

* fix: fix the calculation of gradients of coordnum with pairlist

* Update deprecated Actions version

* Remove CentOS 7 from backend tests

* test: use NAMD 3.0 branch for CI

* test: disable MPI build and run "make depends"

* test: use the correct path of the multicore binary to run tests

* Update NAMD reference output for gradient fixes

* Update test reference data for volmap-based CVs for newer compilers

There are small divergences in numerical results between GCC 4.8 and GCC 11,
which are most likely due to the single-precision math in GridForces.

Updating these tests allows to use a toolchain from a currently maintained OS

* test: add new tests for orientationAngle and orientation

* test: update the tests after fixing the gradients

The following tests are updated:
001_orientation-fitgroup_harmonic-ori-fixed
001_orientationangle_harmonic-fixed

---------

Co-authored-by: Giacomo Fiorin <giacomo.fiorin@gmail.com>
Co-authored-by: Jérôme Hénin <heninj@gmail.com>